Flight Miles Calculator
This program calculates the miles flown for the flights on Phoenix Air, which flies from Phoenix to Los Angeles, San Diego, and Las Vegas. The program asks the user to enter the destination airport code. If the airport code is not valid, the program presents an error message to the user and exits the application. Otherwise, the program calculates the distance the user is flying. The user is asked to specify what their frequent flier status is. Bonus miles are awarded based on the frequent flier status.