Class Quiz Score Management
The application manages the quiz scores for a class of 3 students. The user will specify how
many quizzes each students has. All students will have the same number of quiz scores. The
program prompts the user to enter the name of each student as well as the quiz scores for
each student. It then presents the user with a menu system, which shows a set of possible choices:
1) Calculate Class Average.
2) Calculate Student Average
3) Calculate Quiz Average
The program prompts the user to enter a number corresponding with one of the choices, or
the letter x to exit. As long as the user doesn’t press x, the program continues to show the
same menu to the user.