Student Grade System
This program stores a list of 5 names and corresponding scores and is able to look up scores by name. It utilizes two parallel arrays to store the names and integers, but keeps the order of the names and scores synchronized i.e. if John’s name is in index 1, then John’s score should also be in index 1 in a different array.