Library Management
The library maintains a list of electronic devices. Each device has a Stock Keeping Unit (SKU) code, which
is an alphanumeric code up to 6 letters in length. Each device also has a name. This program
keeps track of whether each device is available or not. Devices are checked out already are considered
“not available”, whereas devices that are “available” can be checked out. A main menu presents the user
with options. Users are able to see a list of devices, add a new device, edit an existing device,
search for a device by name, and check out a device. The main menu asks which option the user
would like to select and then performs that action. After completion, the user is brought back to the
main menu, where they can choose to pick another option or cleanly exit the application.