Loan Manager
The Loan class stores data about a loan, such as interest rate, period of the loan, and the amount of the loan. The Loan class is able to tell what the monthly payment and total payment should be based on the interest rate, period of loan, and loan amount. The main application then asks the user for details for a loan and allows the user to maintain multiple loans. The application maintains an array of Loan objects for this purpose. Finally, the application presents the user the total payment that will be due for all loans combined as well as prints out a list of all loans.